+++ CHANGELOG.md
@@ -6,12 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- Added FreeModule to SharpSploit.Execution.ManualMap.Map (credit @TheWover)
+- Added option to CallMappedDllModuleExport to not invoke EntryPoint (credit @TheWover)
 - Added SharpSploit.Evasion.ETW class, ETWEventWrite patch (credit @salu90)
 - Added SharpSploit.Execution.Injection.VirtualAllocAllocationTechnique (credit @aus)
 - Added SharpSploit.Enumeration.Clipboard clipboard monitor (credit @shellfarmer)
 
+### Changed
+- Update Mimikatz binaries in embedded resources
+
 ### Fixed
 - Fixed SharpSploit.Enumeration.Keylogger HookProc from being garbage collected, fixed handling virtual packets
+- Fixed rewriting IAT of modules with no import table (credit @TheWover)
 
 ## [v1.6] - 2020-06-08
 ### Added